requirements.txt
python-multipart
python-dotenv
semantic-kernel
python-a2a
fastapi
uvicorn
a2a-sdk
